+++ comics/accounts/tests.py
@@ -17,28 +17,22 @@ class LoginTest(TestCase):
 
         self.assertEquals(response.status_code, 302)
         self.assertEquals(
-            response['Location'], '/account/login/?next=/')
+            response['Location'], '/accounts/login/?next=/')
 
     def test_login_page_includes_email_and_password_fields(self):
-        response = self.client.get('/account/login/')
+        response = self.client.get('/accounts/login/')
 
         self.assertEquals(response.status_code, 200)
-        self.assertIn('Email', response.content)
+        self.assertIn('E-mail', response.content)
         self.assertIn('Password', response.content)
-
-    def test_successful_login_redirects_to_front_page(self):
-        response = self.client.post(
-            '/account/login/',
-            {'email': 'alice@example.com', 'password': 'secret'})
-
-        self.assertEquals(response.status_code, 302)
-        self.assertEquals(response['Location'], '/')
 
     def test_failed_login_shows_error_on_login_page(self):
         response = self.client.post(
-            '/account/login/',
-            {'email': 'alice@example.com', 'password': 'wrong'})
+            '/accounts/login/',
+            {'login': 'alice@example.com', 'password': 'wrong'})
 
         self.assertEquals(response.status_code, 200)
         self.assertIn(
-            'Please enter a correct username and password.', response.content)
+            'The e-mail address and/or password you '
+            'specified are not correct.',
+            response.content)